refactor: remove AbstractGeoTable and implement GeoTable directly
