+++ giraffe/src/components/geo/processing/GeoTable.ts
@@ -3,6 +3,11 @@ export interface LatLon {
   lat: number
 }
 
+export enum CoordinateEncoding {
+  GEO_HASH,
+  FIELDS,
+}
+
 export interface Track extends Array<number[]> {}
 
 export interface GeoTable {
