+++ tsconfig.json
@@ -11,6 +11,7 @@
     "moduleResolution": "node",
     "esModuleInterop": true,
     "noImplicitAny": false,
+    "noImplicitThis": false,
     "noUnusedParameters": true,
     "noUnusedLocals": true,
     "strictNullChecks": false,
